text-summarization/requirements.txt
pyyaml
grpcio==1.16.0